+++ Sources/Rapid/Conformances/PartialRangeFromConformances.swift
@@ -1,4 +1,4 @@
-// EquatableRange.swift
+// PartialRangeFromConformances.swift
 // Copyright © 2023 Kaleb A. Ascevich
 //
 // This package is free software: you can redistribute it and/or modify it
@@ -36,46 +36,28 @@ extension PartialRangeFrom: Equatable {
     }
 }
 
-extension PartialRangeThrough: Equatable {
-    /// Returns a Boolean value indicating whether two partial ranges are
-    /// equal.
-    ///
-    /// Two partial ranges are equal when they have the same upper bounds.
-    ///
-    ///     let x = ...15
-    ///     print(x == ...15)
-    ///     // Prints "true"
-    ///     print(x == ...20)
-    ///     // Prints "false"
-    ///
-    /// - Parameters:
-    ///   - lhs: A range to compare.
-    ///   - rhs: Another range to compare.
-    ///
-    /// - Returns: Whether the two ranges are equal.
-    public static func == (lhs: Self, rhs: Self) -> Bool {
-        lhs.upperBound == rhs.upperBound
+extension PartialRangeFrom: Hashable where Bound: Hashable {
+    public func hash(into hasher: inout Hasher) {
+        hasher.combine(lowerBound)
     }
 }
 
-extension PartialRangeUpTo: Equatable {
-    /// Returns a Boolean value indicating whether two partial ranges are
-    /// equal.
-    ///
-    /// Two partial ranges are equal when they have the same upper bounds.
-    ///
-    ///     let x = ..<15
-    ///     print(x == ..<15)
-    ///     // Prints "true"
-    ///     print(x == ..<20)
-    ///     // Prints "false"
-    ///
-    /// - Parameters:
-    ///   - lhs: A range to compare.
-    ///   - rhs: Another range to compare.
-    ///
-    /// - Returns: Whether the two ranges are equal.
-    public static func == (lhs: Self, rhs: Self) -> Bool {
-        lhs.upperBound == rhs.upperBound
+extension PartialRangeFrom: CustomStringConvertible {
+    /// A textual representation of the range.
+    public var description: String {
+        "\(lowerBound)..."
+    }
+}
+
+extension PartialRangeFrom: CustomDebugStringConvertible {
+    /// A textual representation of the range, suitable for debugging.
+    public var debugDescription: String {
+        "PartialRangeFrom(\(String(reflecting: lowerBound))...)"
+    }
+}
+
+extension PartialRangeFrom: CustomReflectable {
+    public var customMirror: Mirror {
+        Mirror(self, children: ["lowerBound": lowerBound])
     }
 }